Fix clang warning. (CaloUtils-01-00-49)
        * Tagging CaloUtils-01-00-49.
	* test/CaloLayerCalculator_test.cxx (Testcell): Fix type of energy
	(fix clang warning).

2014-01-28 Fabian Spettel <fspettel@mppmu.mpg.de>
        * Introduced classify and weight negative clusters as well, off by default
        * Tagging CaloUtils-01-00-48

+++ Calorimeter/CaloUtils/test/CaloLayerCalculator_test.cxx
@@ -57,7 +57,7 @@ struct Testcell
 {
   int ieta;
   int iphi;
-  float energy;
+  double energy;
 };
 
 Testcell clust1[] = {
